// src/types.ts
export interface Announcement {
  id: number;
  title: string;
  date: string;
  content: string;
}

export interface Event {
  id: number;
  title: string;
  date: string;
  location: string;
  description: string;
}

export interface Achievement {
  id: number;
  title: string;
  description: string;
}